fuck this(still unfinished)

+++ src/main/java/io/github/mg138/tsbook/entities/effect/data/EntityStatusEffect.kt
@@ -2,6 +2,6 @@ package io.github.mg138.tsbook.entities.effect.data
 
 import org.bukkit.entity.Entity
 
-class EntityStatusEffect(val target: Entity, type: StatusEffectType, power: Double, ticks: Int) {
+class EntityStatusEffect(val target: Entity, type: StatusEffectType, power: Double, ticks: Long) {
     val statusEffect: StatusEffect = StatusEffect(type, power, ticks)
 }
